requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
discord.py==2.6.4
frozenlist==1.8.0
idna==3.11
mcrcon==0.7.0
multidict==6.7.0
packaging==25.0
propcache==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requirements-parser==0.13.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
